Added support for multiple targets
