+++ generating/wat64.py
@@ -1,0 +1,6 @@
+from typing import List
+from parsing.op import *
+from state import *
+
+def generate_wat64(ops: List[Op]):
+    cont_assert(False, "Not implemented")
